Updates for workshop (#14)
* Improve env and connection handling and logging

* Better logging for config and connection
* Re-order clicks.service.js for demo (readability)
* Always try to create the table if it does not exist

* Organize/split the readme for DBaaS vs OpenShift

Reorganized so there is DB on OpenShift or IBM Cloud and then
start app on OpenShift or locally. Filled out more of the setup.

* History API and page to show users clickstream history

Adding history API/page to show how data can be used.
Raw for now.

* Log time spent on page (category)

Not in DB (yet) just logging, but added the ability
to track time spent on page. Just the category pages
for first test.

* Cleanup unused/commented-out lines in history page

* More AWS app config screenshots and more in sync w/ instructions

The screenshot wasn't quite what the instructions said.
Also, there were a lot of instructions w/ minimal screenshots
which might have been good enough, but was really confusing if
you went looking for screen elements that were not pictured.
So more screenshots were added.

* Add red boxes on the AWS app deploy images